Updated shell scripts

+++ clean.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 clear
-rm bin/*
+
+if [ -d "bin" ]; then
+  rm bin/*
+fi
+
 cd src
 find . -type f -name '*.o' -delete
 exit 0
